lab10 2.0
Ολοκλήρωση κώδικα

Draft αναφορά

+++ 10/code/float/main.c
@@ -1,7 +1,7 @@
 /*
- * long.c
+ * float.c
  *
- * Example program for 3x3 matrix multiplication with long data.
+ * Example program for 3x3 matrix multiplication with float data.
  * It's purpose is to show the time difference when doing the same calculations with long and float format on an avr micro controller.
  *
  * Author : Emmanouil Petrakos
@@ -9,24 +9,24 @@
  * Developed with AtmelStudio 7.0.129
  */ 
 
-void matrix_3x3_multiplication( long input1[3][3] , long input2[3][3] , long output[3][3] );
+void matrix_3x3_multiplication( float input1[3][3] , float input2[3][3] , float output[3][3] );
 
 
 // Declare and initialize arrays.
 // Globals so they don't get optimized away.
-long matrix1[3][3] =
+float matrix1[3][3] =
 {
 	{ 1 , 2 , 3 } ,
 	{ 4 , 5 , 6 } ,
 	{ 7 , 8 , 9 }
 };
-long matrix2[3][3] =
+float matrix2[3][3] =
 {
 	{ 1 , 2 , 3 } ,
 	{ 4 , 5 , 6 } ,
 	{ 7 , 8 , 9 }
 };
-long matrix3[3][3] =
+float matrix3[3][3] =
 {
 	{ 0 , 0 , 0 } ,
 	{ 0 , 0 , 0 } ,
@@ -55,15 +55,15 @@ int main( void )
 
 /**
  * \brief 
- * Matrix multiplication of two 3x3 arrays of longs.
+ * Matrix multiplication of two 3x3 arrays of floats.
  * 
- * \param inputA pointer to the 3x3 array of longs used as the multiplier.
- * \param inputB pointer to the 3x3 array of longs used as the multiplicand.
- * \param output pointer to the 3x3 array of longs where the product is saved.
+ * \param inputA pointer to the 3x3 array of floats used as the multiplier.
+ * \param inputB pointer to the 3x3 array of floats used as the multiplicand.
+ * \param output pointer to the 3x3 array of floats where the product is saved.
  * 
  * \return void
  */
-void matrix_3x3_multiplication( long inputA[3][3] , long inputB[3][3] , long output[3][3] )
+void matrix_3x3_multiplication( float inputA[3][3] , float inputB[3][3] , float output[3][3] )
 {
 	output[0][0] = inputA[0][0] * inputB[0][0] + inputA[0][1] * inputB[1][0] + inputA[0][2] * inputB[2][0];
 	output[0][1] = inputA[0][0] * inputB[0][1] + inputA[0][1] * inputB[1][1] + inputA[0][2] * inputB[2][1];
